moved all api request to one file
