--- src/features/movieDetails/movieDetailsAPI.js
@@ -1,26 +1,0 @@
-import { apiKey } from "../../common/apiData/apiKey";
-import { apiLink } from "../../common/apiData/apiLink";
-
-export const getMovieDetails = async (id) => {
-  const response = await fetch(
-    `${apiLink}/movie/${id}?api_key=${apiKey}&language=en`
-  );
-
-  if (!response.ok) {
-    new Error(response.statusText);
-  }
-
-  return await response.json();
-};
-
-export const getCredits = async (id) => {
-  const response = await fetch(
-    `${apiLink}/movie/${id}/credits?api_key=${apiKey}&language=en`
-  );
-
-  if (!response.ok) {
-    new Error(response.statusText);
-  }
-
-  return await response.json();
-};
