+++ src/features/movieList/movieListSaga.js
@@ -1,5 +1,9 @@
 import { delay, call, put, takeLatest, select } from "redux-saga/effects";
-import { getGenres, getMovies, getMoviesByQuery } from "./movieListAPI";
+import {
+  getGenres,
+  getMovies,
+  getMoviesByQuery,
+} from "../../common/apiData/apiRequests";
 import { loadingDelay } from "../../common/states/loadingDelay";
 import {
   fetchMovies,
